forcing ints return

+++ labelObjects.py
@@ -35,7 +35,7 @@ def labelObj(imagefile, labelfile, t=20, ms=1000, sl=np.ones((3,3,3))):
     else:
       header1['type'] = 'uint8'
       nrrd.write(labelfile, np.uint8(data), options=header1)
-    return np.unique(data)
+    return np.int(np.unique(data))
 
 def cutObj(imagefile, labelfile, labels=None):
     if labels is None:
@@ -77,7 +77,7 @@ def cropObj(imagefile, labelfile, labels=None):
         mask[data2==i] = 0
 
       data1[mask] = 0
-      
+
       v=np.max(data1)
       print "Saving result over " + imagefile
       header1['encoding'] = 'gzip'
